third_party/WebKit/LayoutTests/fast/canvas/canvas-fillStyle-no-quirks-parsing.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>
setup(function() {
  window.ctx = document.createElement('canvas').getContext('2d');
});

test(function() {
  ctx.fillStyle = "#008000";
  ctx.fillStyle = "800000";
  assert_equals(ctx.fillStyle, "#008000");
}, 'fillStyle should not accept quirks mode hex colors.');
</script>
